[Pipeline] Fix sudo tests, add ability to test pipeline locally
sudo tests broke in 837289fd45fc9bfa702d9398ef10c3895849e60f.

- This commit fixes sudo tests to run again

- Commit also adds ability to test pipeline locally, as it turns out any
  change to the docker files are difficult to test on Gitlab as you need
  to construct a new `registry.gitlab.com/yottadb/db/ydbtest` image and
  push it for the pipeline to use it. Now, you just have to change a
  file locally and run the supplied command in the README. It will run
  the tests that are affected by that file.

Sample sudo pipelines:

- https://gitlab.com/shabiel/YDBTest/-/jobs/8265973998
- https://gitlab.com/shabiel/YDBTest/-/jobs/8265973999

Various changes:
- `Dockerfile`: Add libcurl (needed now as the GUI uses it; sudo tests
  install the GUI); move final version build to the end so that we keep
  the files that sudo uses behind.
- Rest of the files have comments that describe the changes.

+++ docker/build_and_install_yottadb.csh
@@ -140,7 +140,11 @@ endif
 mkdir -p $gtm_root/$verno/$dbgpro/obj
 find . -name '*.a' -exec cp {} $gtm_root/$verno/$dbgpro/obj \;
 mkdir -p $gtm_root/$verno/tools
-rm -rf $source_dir/$dbgpro
+
+# Sudo tests require `ydbinstall` that is part of the intermediate build process (yuck); so leave the build directory around for master commit
+if ( "$git_tag" != "master" ) then
+	rm -rf $source_dir/$dbgpro
+endif
 
 cd $source_dir
 cp sr_unix/*.awk $gtm_root/$verno/tools
